Moved NotesPageTopBar to folder /NotesPageTopBar
