+++ src/components/NotesPage/NotesPageTopBar/NotesPageTopBarAction.js
@@ -1,0 +1,5 @@
+export const LOGOUT_ACTION = 'LOGOUT_ACTION';
+
+export const logoutAction = () => ({
+  type: LOGOUT_ACTION,
+});
