+++ src/sagas.js
@@ -8,7 +8,7 @@ import { all } from 'redux-saga/effects';
 
 import { appInit } from 'components/App/AppSagas';
 import { loginInit } from 'components/LoginPage/LoginPageSagas';
-import { logoutInit } from 'components/NotesPage/NotesPageTopBarMenuSagas';
+import { logoutInit } from 'components/NotesPage/NotesPageTopBar/NotesPageTopBarSagas';
 import { noteDetailsInit } from 'components/NoteDetails/NoteDetailsSagas';
 import { noteMindMapInit } from 'components/NotesMindMap/NotesMindMapSagas';
 import { searchPageInit } from 'components/SearchPage/SearchPageSagas';
